# projects/expense-portal/src/openapi.yaml
openapi: "3.1.0"
info:
  title: Expense Management Portal
  description: Employee Expense Management Portal API — Acme Corporation
  version: "0.1.0"
  contact:
    name: Platform Engineering
servers:
  - url: https://expenses.acme.com
    description: Production (via Azure API Management)
paths:
...
  /v1/expenses/reports/{report_id}/line-items/{item_id}/receipt:
    post:
      operationId: uploadReceipt
      tags: [receipts]
      summary: Upload a receipt image
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Receipt uploaded, OCR processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptUploadOut"
  /v1/receipts/{receipt_id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Download receipt (SAS URL redirect)
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "302":
          description: Redirect to blob SAS URL
  /v1/expenses/ocr-status/{task_id}:
    get:
      operationId: getOcrStatus
      tags: [receipts]
      summary: Poll OCR processing status
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OCR status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrStatusOut"
...
  /v1/reports/finance:
    get:
      operationId: financeReport
      tags: [reports]
      summary: Finance reporting dashboard
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [monthly, quarterly, yearly]
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Finance report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceReportOut"
  /v1/reports/manager:
    get:
      operationId: managerReport
      tags: [reports]
      summary: Manager team spend dashboard
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Manager report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerReportOut"
...
components:
  schemas:
...
    ReceiptUploadOut:
      type: object
      properties:
        receipt_url:
          type: string
        ocr_status:
          type: string
        ocr_task_id:
          type: string
    OcrStatusOut:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        extracted_fields:
          type: object
          nullable: true
...
    FinanceReportOut:
      type: object
      properties:
        summary:
          $ref: "#/components/schemas/FinanceSummary"
        by_cost_center:
          type: array
          items:
            type: object
        by_category:
          type: array
          items:
            type: object
        by_status:
          type: array
          items:
            type: object
    FinanceSummary:
      type: object
      properties:
        total_amount:
          type: number
        report_count:
          type: integer
        average_amount:
          type: number
        period:
          type: string
    ManagerReportOut:
      type: object
      properties:
        cost_center:
          type: string
        budget:
          type: number
        total_submitted:
          type: number
        total_approved:
          type: number
        remaining_budget:
          type: number
...